refactor update_metadata
